int16_t

+++ src/software/test.c
@@ -33,9 +33,9 @@ void print_value(void) {
 }
 
 /* Set the background color */
-int* get_value(void)
+int16_t* get_value(void)
 {
-  int* value_local;
+  int16_t* value_local;
   if (ioctl(cnn_fd, CNN_READ_VAL, value_local)) {
       perror("ioctl(CNN_READ_VAL) failed");
       return -1;
